+++ src/frontend/src/views/index.ts
@@ -1,4 +1,5 @@
 // Export all views from a single file for easier imports
 export * from "./CounterView";
 export * from "./GreetingView";
+export * from "./HomeView";
 export * from "./LlmPromptView";